UI for categories
Created the UI for a couple of other categories
