+++ src/pages/business.js
@@ -1,11 +1,11 @@
 
 import React from 'react';
 import './App.css';
-import './title'
-import SimpleSlide from './title';
-import NewsBox from './news_box';
+import './assets/title'
+import SimpleSlide from './assets/title';
+import NewsBox from './assets/news_box';
 import './index.css';
-import Background from './Background.mp4';
+import Background from './assets/Background.mp4';
 
 function Business() {
   return (
@@ -16,7 +16,7 @@ function Business() {
     </video>
   </div>
     <div>
-      <SimpleSlide></SimpleSlide>
+      <SimpleSlide title={'Business'} titleColor = {'White'}></SimpleSlide>
       <NewsBox></NewsBox>
     </div>
   </div>
